Refactored the function translating WebSOC times to 24-hour time into calendarizeHelpers
